### 2023-01-11 #5068 @editorialbot
[PRE REVIEW]: MolPainter: A Tool for Painting and Solvating Layered Molecular Systems
**Submitting author:** @gpantel (<a href="http://orcid.org/0000-0002-4373-1677">George Pantelopulos</a>)
**Repository:** https://github.com/gpantel/MolPainter/
**Branch with paper.md** (empty if default branch): paper_joss
**Version:** v1.1.0
