### 2023-01-11 #5068 @gpantel
Thank you kindly for beginning this process Kyle Niemeyer!

Of the users who have agreed to be reviewers for JoSS, the following would be suitable
rosecers
HubLot
bmansel
corettialessandro
lscalfi
arose
khinsen
malramsay64
govarguz
tonigi/giorginolab
mabraham
KEIPERTK
