### 2023-02-03 #5128 @editorialbot
[REVIEW]: MolPainter: A Tool for Painting and Solvating Layered Molecular Systems
**Submitting author:** @gpantel (<a href="http://orcid.org/0000-0002-4373-1677">George Pantelopulos</a>)
**Repository:** https://github.com/gpantel/MolPainter/
**Branch with paper.md** (empty if default branch): paper_joss
**Version:** 1.1.5
**Archive:** https://doi.org/10.5281/zenodo.7846421
