### 2023-02-06 #5128 @jgostick
Hi @HubLot and @arkajitmandal.  This review is now officially started.  Reviewing for JOSS is pretty straight-forward....you just create a checklist for yourself using the command ``@editorialbot generate my checklist``, then work your way through the items.
