### 2023-02-10 gpantel/MolPainter#15 @HubLot
[JOSS Review] Automated tests
This issue is part of the JOSS review process available here: https://github.com/openjournals/joss-reviews/issues/5128

---

According to [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#tests):

      Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.

As far as I can tell, MolPainter does not contain any automated or manual tests. I strongly encourage to include tests, at least for the drawing part (to make sure the system is correctly created with lipids and/or solute) and for Molsolvator (to make sure the solvation process is  running smoothly).

The [pytest](https://docs.pytest.org/en/7.2.x/) testing framework is a good tool to implement those kind of tests. Additionally, once they are implemented, it would be useful to run those tests automatically through a continuous integration service like [Github Actions](https://docs.github.com/en/actions).
