### 2023-02-10 gpantel/MolPainter#16 @HubLot
[JOSS Review] Add contribution guidelines
This issue is part of the JOSS review process available here: https://github.com/openjournals/joss-reviews/issues/5128

---

According to [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines):

      There should be clear guidelines for third-parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

A guide for contributing should be added to the repository, see [here](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) as a starting point.
