### 2023-02-10 gpantel/MolPainter#17 @HubLot
[JOSS Review] Installation instructions & version number
This issue is part of the JOSS review process available here: https://github.com/openjournals/joss-reviews/issues/5128

---

### Installation instructions
The installation process is quite straightforward thanks to `pip`.  I managed to install it on different systems. However, the README section is minimalist.
I would recommend that the authors provide a list of supported platforms and a list of the dependencies of the software.

## Version number

Currently, there is no easy way to find out which version of the software I'm using when installing it from `pip` (besides using the `pip freeze` command). It's not available in the software commands or in the README.

I would recommend that the authors :

 -  provide a mechanism to output the version on the software side (this could be a new button 'about' in the MolPainter part and a `-v` option in MolSolvator).
 - add the pypi badge to the README to display the latest Pypi version available.
 - optionally, use `git tags` to create version archives of the software and provide permanent links between the pypi version and the Github repository.
