### 2023-02-13 gpantel/MolPainter#18 @HubLot
[JOSS Review] Documentation improvements 
This issue is part of the JOSS review process available here: https://github.com/openjournals/joss-reviews/issues/5128

---

The documentation is a bit scattered across different supports:
   - README.md in the repo
   - Tutorial.md in another location in the repo
   - the paper.

Furthermore, **the paper is too long** according to [JOSS Submission guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) (2962 words for the paper against 250-1000 words suggested) and contains too much information not found elsewhere in the repo.

This makes difficult for the user to easily find the correct answer to his/her questions (for example, how to define weight for the blend? What is the z-position for the layer?)

The documentation could be greatly improved by:

  - shrinking the size of the paper by removing major parts of documentation.
  - providing a single point of entry for the documentation. This could be the current README with a table of contents and links to various documentation files. This could also be files formatted in a markup language (using tools like [Sphinx](https://www.sphinx-doc.org/en/master/) or [mkdocs](https://www.mkdocs.org/)) and hosted on a service like [Read the docs](https://readthedocs.org/).
  - explaining, in a single location, each button and each option of those buttons (and not only in the tutorial).
