### 2023-02-14 gpantel/MolPainter#19 @HubLot
[JOSS Review] MolPainter tool minor improvements
This issue is part of the JOSS review process available here: https://github.com/openjournals/joss-reviews/issues/5128

---

Overall, the GUI based Molpainter tool is a good piece of work. The GUI is stable and the interaction is smoothed.
I have only minor suggestions:
   - add a "about" button with version number, link to the documentation and to the GitHub repository. This will allow users to easily find useful information,= and documentation about the software.
   - Rename "Open Painting" by "Load Painting". It will be clearer for the users that this button is for "loading" a configuration file and not "opening" a random file.
   - When using the pencil painting tool, is it possible to paint one cell by just clicking on it and not clicking and dragging? This will make easier to paint tiny cells.
   - The % of molecules printed on the left is independent of the number of cells labeled "empty". Wouldn't it be better that the sum of all percentage represent the total surface of the layer and thus, taking into account the *empty* cells ?
