### 2023-02-14 #5128 @HubLot
Hi @jgostick and @gpantel, I have started my review and overall, I think it's a good software and clearly fulfils a need. The ability to create complex solvated lipid mixtures (or any layered molecular system) by painting is valuable.

However, at this stage, the submission does not fulfil JOSS Guidelines requirements:
   - **Paper**: While the paper contains the correct sections, it is far too long (~3000 words against the 1000 max suggested) and contains valuable information/documentation not available elsewhere. I believe some sections (like the detailed description of MolPainter and Molsolvator input parameters) should be removed and integrated to the software documentation.
   See https://github.com/gpantel/MolPainter/issues/18
   - **Tests**: There appear to be no tests included with the software. I highly recommend the author consider using a testing framework for their software.
   See https://github.com/gpantel/MolPainter/issues/15
   - **Community**: There is no clear set of guidelines for how to contribute or a code of conduct.
   See https://github.com/gpantel/MolPainter/issues/16

I also have some minor issues which are suggestions:
   - install requirements.
   See https://github.com/gpantel/MolPainter/issues/17
   - general improvements of Molpainter tool.
   See https://github.com/gpantel/MolPainter/issues/19
   - general improvements of MolSolvator tool.
   See https://github.com/gpantel/MolPainter/issues/20
   - **tutorial**: the tutorial is well written and explain in details the different stages to produce a fully solvated lipid bilayer and all files needed or generated are present. I have just few comments: the shape of the layer in Figure 3 is not correct compared to the values in Figure 2. In step 3,4 and 7, the button we need to click is not specified. GROMACS version used should be specified.
