### 2023-02-14 #5128 @gpantel
@HubLot Thank you for taking so much time to review our submission and making important suggestions I and @a-a-ron1 find helpful and will make the program more accessible and easier to disseminate.

I'm preparing for a conference and hosting a visiting scientist this week so I'm not able to make all of these changes right away. May the JOSS Guidelines issues be prioritizes to complete for publication, and the remaining issues might be addressed in follow-up update?
