### 2023-02-21 gpantel/MolPainter#21 @arkajitmandal
[JOSS Review]  The text "Layer #" do not show up for the active layer
When first loading the software the text "Layer 1" is not shown.
It shows up when right clicking on the tab,

On closer inspection the issue seems to that the active/on-click color of text for buttons is chosen to be white, thus making it disappear if the button background is white.

To me, an easy fix it to swap the color white for the button/tab background for Layers to something dark.
