### 2023-02-21 #5128 @arkajitmandal
Hi @jgostick and @gpantel, I have checked the software and used it following the documentation provided by the authors. The software overall is very well made and should be very useful in the field  of computational chemistry/biophysics.

While this work is overall very nice, it does require a bit of clean up (addressing few issues) for it to be recommended for publication. I noticed many of them have been already been mentioned by the other reviewer (so I wont make a duplicate issue).  The one additional issue, I think the authors should address is:

* Seems the text for active/pressed buttons have a default color white. As a result, sometimes they do not show up when the buttons background is also white. I suggest that the authors change the background color or the text color to something bit darker to avoid this issue.
See "https://github.com/gpantel/MolPainter/issues/21#issue-1592676584"
