+++ CONTRIBUTING.md
@@ -1,0 +1,25 @@
+# Contributing to MolPainter
+
+We thank you for using our project and we'd love to hear from you.
+
+# Asking for Help
+
+You can email us at 
+
+* Open a discussion on GitHub
+* Email us
+
+# Reporting Issues
+
+* Open an issue
+* If MolPainter or MolSolvator experience an error please provide
+  ** The input PDB files you're using
+  ** The operating system you're working on
+  ** The parameters used for building the system
+
+# Making Changes
+
+# Let us promote your work using MolPainter!
+
+* Please reach out to us an let us know about the cool work you've done!
+* We would like to promote interesting systems constructed using MolPainter
