### 2023-03-18 54a56e6
Documentation draft
Pulling documentation out of paper draft in response to reviewer comments
