+++ tutorial/DOCUMENTATION.md
@@ -1,0 +1,75 @@
+# MolPainter and MolSolvator functions
+
+MolPainter is a novel GUI tool for constructing layered molecular systems represented using PDB files. The complementary command line MolSolvator is a simple lattice-based solvation tool for use in solvating molecular systems constructed with MolPainter.
+
+The purpose of MolPainter is to enable the construction of complex multilayered molecular systems through fine control over the position of individual molecules, using the allegory of painting as a means by which to select these positions in cartesian coordinates.
+
+## MolPainter objects and functions
+![MolPainter components and functions. (A) *n* layers defined at assigned z-axial coordinates in angstrom units. (B) the grid used to define the xy-plane position of each lipid in terms of a cell spacing in angstrom units and number of cells in the x- and y-dimension. (C) The molecules are defined by PDB files supplied by the user. (D) Anchor points, defined in PDB files using B-factors of 1.00, define where the centroid of all anchor points z-axially align into the grid layer pained upon. (E) The canvas holds labels of each molecule painted onto xy-positions in each grid on each layer via various painting tools, such as the spray can. (G) Solutes can be inserted to the canvas, such as the C99 protein monomer shown here,[@Pantelopulos2022] obfuscating cells across multiple grid layers to prevent clashes with painted molecules.\label{fig:MolPainterFigure}](images/MolPainterObjectsAndFunctions.png)
+
+### Layers
+Systems in MolPainter are represented by one or more layers. Each layer is represented by a grid defined at a given position on the Z-axis (Figure \autoref{fig:MolPainterFigure}.A). This enables the creation of layered systems as well as simple three-dimensional structures.
+
+### The Grid
+
+Each layer is characterized by a grid which serves as a "canvas" for (I) precise placement (painting) of molecules in space relative to other molecules and (II) the precise counting of molecules while painting the grid. Each cell in the grid is a square cell with a length given in angstroms, defined as the cell spacing (Figure \autoref{fig:MolPainterFigure}.B). The width (X-axis) and the height (Y-axis) of the grid is defined by the number of cells in each dimension. It should be noted that the grid is only for relative spatial placement and that there is no representation of the underlying geometry of the molecules.
+
+### Molecules
+Molecules are the primary input of the application and are configurable both in terms of how they are represented in the painting process and how they are generated in the exported system. For painting purposes, each molecule is assigned a name and a color, which serve to identify it among other molecules. Within each layer, the number and mol% of each molecule is tracked with a counter. This enables construction of a system with a desired amount or ratio of different molecule types.
+
+For system construction (export) purposes, each molecule is associated with a PDB file and a set of optional rotational transformations to be applied during export. The PDB file is read during export and its entire contents are replicated for each instance (painted cell) of that molecule in the painting (Figure \autoref{fig:MolPainterFigure}.C). The X and Y coordinates of the painted cell are determined by its placement in the grid; its Z coordinate is determined by the Z coordinate of the containing layer. When exporting, the X, Y, and Z coordinates of each molecule are translationally transformed to align with the X, Y, and Z coordinates of each painted cell.
+
+The coordinates of each molecule can be transformed by rotating 180$^\circ$ over the X-axis. This transformation is uniformly applied to each painted instance of the molecule type. The molecules can also be randomly rotated around the Z-axis. This transformation is applied with a new random value for each painted instance of the molecule type. The transformations can be used for such purposes as flipping lipids to form bilayers and randomly rotating the orientations of proteins in the XY-plane.
+
[…]
