+++ README.md
@@ -3,12 +3,8 @@
 ## Table of Contents
 - [Introducing MolPainter and MolSolvator](#introducing-molpainter-and-molsolvator)
 - [Installation](#installation)
-- [MolPainter Usage](#molpainter-usage)
-  * [Buttons](#buttons)
-  * [Molecule palette](#molecule-palette)
-  * [Layers](#layers)
-  * [PDB files for molecules](#pdb-files-for-molecules)
-- [MolSolvator Usage](#molsolvator-usage)
+- [MolPainter Usage](#molpainter-documentation)
+- [MolPainter Tutorial](#molpainter-tutorial)
 - [Give it a try!](#give-it-a-try)
 
 ## Introducing MolPainter and MolSolvator
@@ -36,73 +32,13 @@ molpainter
 molsolvator -i input.toml
 ```
 
-## MolPainter Usage
-### Buttons
-* Topbar buttons
-    - **New** resets everything in MolPainter.
-    - **Open Painting** loads a saved MolPainter state so you can continue changing your painting.
-    - **Save Painting** saves a MolPainter state for later! This is a JSON-format file.
-    - **Export System** constructes your painted system in PDB format. You can construct your system on square lattices or hexagonal lattices.
-    - **Import Solute** *NEW IN 1.1!!* inserts a solute (PDBfile) to the MolPainter canvases, obstructing cells occupied by the solute coordinates and outputting the solute with the painting when exported.
-    - **New Molecule** loads a PDB file corresponding to a molecule you wish to paint with. You can also set all painted copies of the molecule to be randomly rotated around the z-axis or be flipped 180˚ over the x-axis (like for lipid bilayers).
-    - **New Blend** makes a "blend" of molecules. Blends draw molecules onto layer canvases with a weighted probability, running a RNG to determine which of the molecules that compose the blend are painted given their weighted probability.
-    - **New Layer** makes new layers to paint on, defined by a z-axial position.
-    - **Grid Size** sets the lattice width - the number of cells in each layer along the x-axis, the lattice height - the number of cells in each layer along the y-axis, the lattice spacing - the length of *each* cell, and the major gridline spacing, used for visual guidance when painting. 
+## Documentation
 
-* Toolbox buttons
-    - **Magnifying glass -** zooms out from the lattice view.
-    - **Magnifying glass +** zooms in on the lattice view.
-    - **Pencil** enables the pencil drawing tool. Click and hold to draw selected molecule/blend.
-    - **Square** enables the square selection tool. Click position to start drawing a rectilinear shape, hold, move, and release at position to fill selected molecule/blend into the rectilinear shape. *Can also be used via a ctrl+click shortcut*
-    - **Spray can** sprays selected molecule/blend into an area. *Can also be used via a shift+click shortcut*
-    - **Spray radius** adjust the radius of the spray can.
+Descriptions of the objects and functions of MolPainter and MolSolvator are available [here](/tutorial/DOCUMENTATION.md)
 
-### Molecule palette
-The molecule palette, located to the lower left of MolPainter, contains a list of all molecules and blends loaded via the *New Molecule* and *New Blend* buttons, in order. The *Empty* molecule is always present, and is essentially an "eraser".
+## Tutorial
 
-When you want to paint with a molecule or blend in the palette, click on the radio button. You can *edit*, *copy*, or *delete* any molecule via a drop-down menu that appears when right-clicked.
-
-### Layers
-The canvas layers to which molecules and blends are painted, created using *New Layer*, can be selected for painting by clicking on the corresponding layer button, which can be named. Layer names and z-axial positions can also be edited via a drop-down menu that appears when the layer button is right-clicked.
-
-### PDB files for molecules
[…]
