+++ joss_manuscript/paper.bib
@@ -475,3 +475,27 @@ url = {http://www.ncbi.nlm.nih.gov/entrez/query.fcgi?cmd=Retrieve&db=PubMed&dopt
 volume = {82},
 year = {2002}
 }
+@article{MacDermaid2020,
+title = {Coexistence of Lipid Phases Stabilizes Interstitial Water in the Outer Layer of Mammalian Skin},
+journal = {Biophysical Journal},
+volume = {118},
+number = {7},
+pages = {1588-1601},
+year = {2020},
+issn = {0006-3495},
+doi = {https://doi.org/10.1016/j.bpj.2020.01.044},
+url = {https://www.sciencedirect.com/science/article/pii/S0006349520301211},
+author = {Christopher M. MacDermaid and Kyle Wm. Hall and Russell H. DeVane and Michael L. Klein and Giacomo Fiorin},
+abstract = {The lipid matrix in the outer layer of mammalian skin, the stratum corneum, has been previously investigated by multiple biophysical techniques aimed at identifying hydrophilic and lipophilic pathways of permeation. Although consensus is developing over the microscopic structure of the lipid matrix, no molecular-resolution model describes the permeability of all chemical species simultaneously. Using molecular dynamics simulations of a model mixture of skin lipids, the self-assembly of the lipid matrix lamellae has been studied. At higher humidity, the resulting lamellar phase is maintained by partitioning excess water into isolated droplets of controlled size and spatial distribution. The droplets may fuse together to form intralamellar water channels, thereby providing a pathway for the permeation of hydrophilic species. These results reconcile competing data on the outer skin’s structure and broaden the scope of molecular-based methods to improve the safety of topical products and to advance transdermal drug delivery.}
+}
+@article{Bouwstra2001,
+  title={Phase behavior of lipid mixtures based on human ceramides: coexistence of crystalline and liquid phases},
+  author={Bouwstra, Joke A and Gooris, Gert S and Dubbelaar, FER and Ponec, Maria},
+  journal={Journal of lipid research},
+  volume={42},
+  number={11},
+  pages={1759--1770},
+  year={2001},
+  publisher={ASBMB}
+}
+
