+++ joss_manuscript/paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 
 # Summary
 
-MolPainter presents the metaphor of painting as an abstraction for the construction of complex layered molecular systems. In particular, systems whose equilibrium states exhibit distinctive visual artifacts can be intentionally crafted through interactive, visually-oriented molecule placement. This is in contrast to non-interactive methods, which allow neither manual placement of molecules nor visual confirmation of desired states. The principal use-case is presented for modelling lipid bilayers with molecular dynamics simulation, though MolPainter may be used to construct any layered molecular system provided appropriate PDB-format files to define molecules.
+MolPainter presents the metaphor of painting as an abstraction for the construction of complex layered molecular systems. In particular, systems whose equilibrium states exhibit distinctive visual artifacts can be intentionally crafted through interactive, visually-oriented molecule placement. This is in contrast to non-interactive methods, which allow neither manual placement of molecules nor visual confirmation of desired states. The principal use-case is presented for modelling lipid bilayers with molecular dynamics simulation, though MolPainter may be used to construct any layered molecular system provided appropriate PDB-format files to define molecules (\autoref{fig:Figure1}).
 
 ![MolPainter uses xy-plane grids at defined z-positions and allows the user to paint user-supplied PDB files into these grids to define a molecular system. MolSolvator solvates these systems after construction.\label{fig:Figure1}](figures/MolPainterGraphic.png)
 
@@ -50,7 +50,7 @@ Along the z-axis, multi-layered lipid systems, such as the long-period-periodici
 
 ![MolPainter components and functions. (A) *n* layers defined at assigned z-axial coordinates in angstrom units. (B) the grid used to define the xy-plane position of each lipid in terms of a cell spacing in angstrom units and number of cells in the x- and y-dimension. (C) The molecules are defined by PDB files supplied by the user. (D) Anchor points, defined in PDB files using B-factors of 1.00, define where the centroid of all anchor points z-axially align into the grid layer pained upon. (E) The canvas holds labels of each molecule painted onto xy-positions in each grid on each layer via various painting tools, such as the spray can. (G) Solutes can be inserted to the canvas, such as the C99 protein monomer shown here,[@Pantelopulos2022] obfuscating cells across multiple grid layers to prevent clashes with painted molecules.\label{fig:Figure2}](figures/MolPainterObjectsAndFunctions.png)
 
-Through the use of user-provided PDB files, MolPainter can be used to paint molecules onto a canvas, a lattice in the xy-plane, aligning each molecule to a specific xy-coordinate and specific z-positions defined for any number of canvases (Figure \autoref{fig:Figure2}). MolPainter also supports the inclusion of a "solute" to the painting, allowing for ease in constructing unique environments defined by complex mixtures of solvents. With MolPainter we include a light companion script, MolSolvator, for the addition of bulk solvent to systems painted using MolPainter.
+Through the use of user-provided PDB files, MolPainter can be used to paint molecules onto a canvas, a lattice in the xy-plane, aligning each molecule to a specific xy-coordinate and specific z-positions defined for any number of canvases (\autoref{fig:Figure2}). MolPainter also supports the inclusion of a "solute" to the painting, allowing for ease in constructing unique environments defined by complex mixtures of solvents. With MolPainter we include a light companion script, MolSolvator, for the addition of bulk solvent to systems painted using MolPainter.
 
 # Acknowledgements
 
