+++ joss_manuscript/paper.bib
@@ -2,7 +2,7 @@
 abstract = {Biological macromolecules function in highly crowded cellular environments. The structure and dynamics of proteins and nucleic acids are well characterized in vitro, but in vivo crowding effects remain unclear. Using molecular dynamics simulations of a comprehensive atomistic model cytoplasm we found that protein-protein interactions may destabilize native protein structures, whereas metabolite interactions may induce more compact states due to electrostatic screening. Protein-protein interactions also resulted in significant variations in reduced macromolecular diffusion under crowded conditions, while metabolites exhibited significant two-dimensional surface diffusion and altered protein-ligand binding that may reduce the effective concentration of metabolites and ligands in vivo. Metabolic enzymes showed weak non-specific association in cellular environments attributed to solvation and entropic effects. These effects are expected to have broad implications for the in vivo functioning of biomolecules. This work is a first step towards physically realistic in silico whole-cell models that connect molecular with cellular biology.},
 author = {Yu, Isseki and Mori, Takaharu and Ando, Tadashi and Harada, Ryuhei and Jung, Jaewoon and Sugita, Yuji and Feig, Michael},
 doi = {10.7554/eLife.19274},
-file = {:Users/gpantel/Library/Application Support/Mendeley Desktop/Downloaded/Yu et al. - 2016 - Biomolecular interactions modulate macromolecular structure and dynamics in atomistic model of a bacterial cytoplasm.pdf:pdf},
+ bacterial cytoplasm.pdf:pdf},
 issn = {2050-084X},
 journal = {eLife},
 month = {nov},
@@ -13,64 +13,66 @@ url = {http://elifesciences.org/lookup/doi/10.7554/eLife.19274},
 volume = {5},
 year = {2016}
 }
-@article{Mauritz2004,
-author = {Mauritz, Kenneth A. and Moore, Robert B.},
-doi = {10.1021/cr0207123},
-file = {:Users/gpantel/Downloads/paper_downloads/cr0207123.pdf:pdf},
-issn = {00092665},
-journal = {Chemical Reviews},
-number = {10},
-pages = {4535--4585},
-pmid = {15669162},
-title = {{State of understanding of Nafion}},
-volume = {104},
-year = {2004}
-}
-@article{Lopez-Haro2014,
-abstract = {Proton exchange membrane fuel cell is one of the most promising zero-emission power sources for automotive or stationary applications. However, their cost and lifetime remain the two major key issues for a widespread commercialization. Consequently, much attention has been devoted to optimizing the membrane/electrode assembly that constitute the fuel cell core. The electrodes consist of carbon black supporting Pt nanoparticles and Nafion as the ionomer binder. Although the ionomer plays a crucial role as ionic conductor through the electrode, little is known about its distribution inside the electrode. Here we report the three-dimensional morphology of the Nafion thin layer surrounding the carbon particles, which is imaged using electron tomography. The analyses reveal that doubling the amount of Nafion in the electrode leads to a twofold increase in its degree of coverage of the carbon, while the thickness of the layer, around 7{\^{a}} €‰nm, is unchanged.},
-author = {Lopez-Haro, M. and Gu{\'{e}}taz, L. and Printemps, T. and Morin, A. and Escribano, S. and Jouneau, P. H. and Bayle-Guillemaud, P. and Chandezon, F. and Gebel, G.},
-doi = {10.1038/ncomms6229},
-file = {:Users/gpantel/Downloads/paper_downloads/ncomms6229.pdf:pdf},
-issn = {20411723},
-journal = {Nature Communications},
-pages = {1--6},
-title = {{Three-dimensional analysis of Nafion layers in fuel cell electrodes}},
-volume = {5},
-year = {2014}
-}
-@article{Wu2014a,
[…]

[…]
