+++ joss_manuscript/paper.bib
@@ -2,7 +2,6 @@
 abstract = {Biological macromolecules function in highly crowded cellular environments. The structure and dynamics of proteins and nucleic acids are well characterized in vitro, but in vivo crowding effects remain unclear. Using molecular dynamics simulations of a comprehensive atomistic model cytoplasm we found that protein-protein interactions may destabilize native protein structures, whereas metabolite interactions may induce more compact states due to electrostatic screening. Protein-protein interactions also resulted in significant variations in reduced macromolecular diffusion under crowded conditions, while metabolites exhibited significant two-dimensional surface diffusion and altered protein-ligand binding that may reduce the effective concentration of metabolites and ligands in vivo. Metabolic enzymes showed weak non-specific association in cellular environments attributed to solvation and entropic effects. These effects are expected to have broad implications for the in vivo functioning of biomolecules. This work is a first step towards physically realistic in silico whole-cell models that connect molecular with cellular biology.},
 author = {Yu, Isseki and Mori, Takaharu and Ando, Tadashi and Harada, Ryuhei and Jung, Jaewoon and Sugita, Yuji and Feig, Michael},
 doi = {10.7554/eLife.19274},
- bacterial cytoplasm.pdf:pdf},
 issn = {2050-084X},
 journal = {eLife},
 month = {nov},
@@ -78,7 +77,6 @@ year = {2019}
 abstract = {For simulations of membranes and membrane proteins, the generation of the lipid bilayer is a critical step in the setup of the system. Membranes comprising multiple components pose a particular challenge, because the relative abundances need to be controlled and the equilibration of the system may take several microseconds. Here we present a comprehensive method for building membrane containing systems, characterized by simplicity and versatility. The program uses preset, coarse-grain lipid templates to build the membrane, and also allows on-the-fly generation of simple lipid types by specifying the headgroup, linker, and lipid tails on the command line. The resulting models can be equilibrated, after which a relaxed atomistic model can be obtained by reverse transformation. For multicomponent membranes, this provides an efficient means for generating equilibrated atomistic models. The method is called insane, an acronym for INSert membrANE. The program has been made available, together with the complementary method for reverse transformation, at http://cgmartini.nl/. This work highlights the key features of insane and presents a survey of properties for a large range of lipids as a start of a computational lipidomics project.},
 author = {Wassenaar, Tsjerk A. and Ing{\'{o}}lfsson, Helgi I. and B{\"{o}}ckmann, Rainer A. and Tieleman, D. Peter and Marrink, Siewert J.},
 doi = {10.1021/acs.jctc.5b00209},
-file = {:Users/gpantel/Downloads/paper_downloads/Wassenaar et al. - 2015 - Computational Lipidomics with insane A Versatile Tool for Generating Custom Membranes for Molecular Simulatio.pdf:pdf},
 isbn = {1549-9618},
 issn = {1549-9618},
 journal = {Journal of Chemical Theory and Computation},
@@ -95,7 +93,6 @@ year = {2015}
[…]
