+++ joss_manuscript/paper.md
@@ -44,7 +44,7 @@ Lipid membranes are believed to exhibit specific, functional clusters of lipids,
 
 The shape of such molecular phase separations and clusters can also be unique. The $\mathrm{L_o}$ phase at high cholesterol concentrations has been proposed to form "threads" of cholesterol, in which each cholesterol has no more and no less than two cholesterol in the first solvation shell, manifesting a maze-like pattern.[@Huang1999; @Miao2002; @Pantelopulos2018a] This has been proposed to either be due to unfavorable penetration of water introduced by loss of protective lipid headgroups or specific cholesterol-cholesterol interactions involving the $\alpha$ and $\beta$ faces of cholesterol.[@Bandara2017] The shape and size of lipid phase separations can also vary significantly, manifesting as stripes or dots in vesicles[@Pantelopulos2017] and molecular simulations with varying system dimensions and line tensions between the $\mathrm{L_o}$ and $\mathrm{L_d}$ phase.[@Kwon2022] 
 
-Along the z-axis, multi-layered lipid systems, such as the long-period-periodicity (LPP) phase of lipids,[@Bouwstra2001] have been an interesting subject of lipid biophysics for their importance in mediating permeation of small molecules through skin. Large scale molecular simulations of model multilamellar skin lipids have been performed to better understand these complex environments.[@MacDermaid2022] MolPainter is uniquely well-suited to construction of systems with complex lateral distributions as well as such multi-layered lipid systems.
+Along the z-axis, multi-layered lipid systems, such as the long-period-periodicity (LPP) phase of lipids,[@Bouwstra2001] have been an interesting subject of lipid biophysics for their importance in mediating permeation of small molecules through skin. Large scale molecular simulations of model multilamellar skin lipids have been performed to better understand these complex environments.[@MacDermaid2020] MolPainter is uniquely well-suited to construction of systems with complex lateral distributions as well as such multi-layered lipid systems.
 
 # Functions
 
