+++ joss_manuscript/paper.bib
@@ -211,6 +211,19 @@ url = {http://aip.scitation.org/doi/10.1063/1.4999709},
 volume = {147},
 year = {2017}
 }
+@article{Pantelopulos2022,
+abstract = {The 99-residue C-terminal domain of amyloid precursor protein (APP-C99), precursor to amyloid beta (A $\beta$ ), is a transmembrane (TM) protein containing intrinsically disordered N- and C-terminal extramembrane domains. Using molecular dynamics (MD) simulations, we show that the structural ensemble of the C99 monomer is best described in terms of thousands of states. The C99 monomer has a propensity to form $\beta$ -strand in the C-terminal extramembrane domain, which explains the slow spin relaxation times observed in paramagnetic probe NMR experiments. Surprisingly, homodimerization of C99 not only narrows the conformational ensemble from thousands to a few states through the formation of metastable $\beta$ -strands in extramembrane domains but also stabilizes extramembrane $\alpha$ -helices. The extramembrane domain structure is observed to dramatically impact the homodimerization motif, resulting in the modification of TM domain conformations. Our study provides an atomic-level structural basis for communication between the extramembrane domains of the C99 protein and TM homodimer formation. This finding could serve as a general model for understanding the influence of disordered extramembrane domains on TM protein structure.},
+author = {Pantelopulos, George A and Matsuoka, Daisuke and Hutchison, James M and Sanders, Charles R and Sugita, Yuji and Straub, John E and Thirumalai, D},
+doi = {10.1073/pnas.2212207119},
+issn = {0027-8424},
+journal = {Proceedings of the National Academy of Sciences},
+month = {dec},
+number = {52},
+title = {{Formation of extramembrane $\beta$ -strands controls dimerization of transmembrane helices in amyloid precursor protein C99}},
+url = {https://pnas.org/doi/10.1073/pnas.2212207119},
+volume = {119},
+year = {2022}
+}
 @article{Kwon2022,
 abstract = {An approach for the efficient simulation of phase-separated lipid bilayers, for use in the calculation of equilibrium free energies of partitioning between lipid domains, is proposed. The methodology exploits restraint potentials and rectangular aspect ratios that enforce lipid phase separation, allowing for the simulation of smaller systems that approximately reproduce bulk behavior. The utility of this approach is demonstrated through the calculation of potentials of mean force for the translation of a transmembrane protein between lipid domains. The impact of the imposed restraints on lipid tail ordering and lipid packing are explored, providing insight into how restraints can best be employed to compute accurate free-energy surfaces. This approach should be useful in the accurate calculation of equilibrium partition coefficients for transmembrane protein partitioning in heterogeneous membranes, providing insight into the thermodynamic driving forces that control this fundamental biophysical phenomenon.},
 author = {Kwon, Seulki and Pantelopulos, George A. and Straub, John E.},
