+++ joss_manuscript/paper.bib
@@ -1,6 +1,4 @@
 @article{Yu2016,
-abstract = {Biological macromolecules function in highly crowded cellular environments. The structure and dynamics of proteins and nucleic acids are well characterized in vitro, but in vivo crowding effects remain unclear. Using molecular dynamics simulations of a comprehensive atomistic model cytoplasm we found that protein-protein interactions may destabilize native protein structures, whereas metabolite interactions may induce more compact states due to electrostatic screening. Protein-protein interactions also resulted in significant variations in reduced macromolecular diffusion under crowded conditions, while metabolites exhibited significant two-dimensional surface diffusion and altered protein-ligand binding that may reduce the effective concentration of metabolites and ligands in vivo. Metabolic enzymes showed weak non-specific association in cellular environments attributed to solvation and entropic effects. These effects are expected to have broad implications for the in vivo functioning of biomolecules. This work is a first step towards physically realistic in silico whole-cell models that connect molecular with cellular biology.},
-author = {Yu, Isseki and Mori, Takaharu and Ando, Tadashi and Harada, Ryuhei and Jung, Jaewoon and Sugita, Yuji and Feig, Michael},
 doi = {10.7554/eLife.19274},
 issn = {2050-084X},
 journal = {eLife},
@@ -13,7 +11,6 @@ volume = {5},
 year = {2016}
 }
 @article{Martinez2009,
-abstract = {Gabedit is a freeware graphical user interface, offering preprocessing and postprocessing adapted (to date) to nine computational chemistry software packages. It includes tools for editing, displaying, analyzing, convert- ing, and animating molecular systems. A conformational search tool is implemented using a molecular mechanics or a semiempirical potential. Input files can be generated for the computational chemistry software supported by Gabe- dit. Some molecular properties of interest are processed directly from the output of the computational chemistry pro- grams; others are calculated by Gabedit before display. Molecular orbitals, electron density, electrostatic potential, nuclear magnetic resonance shielding density, and any other volumetric data properties can be displayed. It can dis- play electronic circular dichroism, UV–visible, infrared, and Raman-computed spectra after a convolution. Gabedit can generate a Povray file for geometry, surfaces, contours, and color-coded planes. Output can be exported to a selection of popular image and vector graphics file formats; the program can also generate a series of pictures for animation. Quantum mechanical electrostatic potentials can be calculated using the partial charges on atoms, or by solving the Poisson equation using the multigrid method. The atoms in molecule charges can also be calculated. Gabedit is platform independent. The code is distributed under free open source X11 style license and is available at http://gabedit.sourceforge.net/.},
 author = {Mart{\'{i}}nez, L. and Andrade, R. and Birgin, E. G. and Mart{\'{i}}nez, J. M.},
 doi = {10.1002/jcc.21224},
 isbn = {1096-987X},
@@ -30,7 +27,6 @@ volume = {30},
 year = {2009}
 }
 @article{VanDerSpoel2005,
[…]
