+++ joss_manuscript/paper.bib
@@ -1,4 +1,5 @@
 @article{Yu2016,
+author = {Yu, Isseki and Mori, Takaharu and Ando, Tadashi and Harada, Ryuhei and Jung, Jaewoon and Sugita, Yuji and Feig, Michael},
 doi = {10.7554/eLife.19274},
 issn = {2050-084X},
 journal = {eLife},
@@ -55,6 +56,7 @@ volume = {29},
 year = {2008}
 }
 @article{Schott-Verdugo2019,
+author = {Schott-Verdugo, Stephan and Gohlke, Holger},
 doi = {10.1021/acs.jcim.9b00269},
 issn = {1549-9596},
 journal = {Journal of Chemical Information and Modeling},
@@ -67,6 +69,7 @@ volume = {59},
 year = {2019}
 }
 @article{Wassenaar2015,
+author = {Wassenaar, Tsjerk A. and Ing{\'{o}}lfsson, Helgi I. and B{\"{o}}ckmann, Rainer A. and Tieleman, D. Peter and Marrink, Siewert J.},
 doi = {10.1021/acs.jctc.5b00209},
 isbn = {1549-9618},
 issn = {1549-9618},
@@ -220,6 +223,7 @@ url = {https://doi.org/10.1016/j.bpj.2022.07.031},
 year = {2022}
 }
 @article{MacDermaid2020,
+author = {Christopher M. MacDermaid and Kyle Wm. Hall and Russell H. DeVane and Michael L. Klein and Giacomo Fiorin},
 title = {Coexistence of Lipid Phases Stabilizes Interstitial Water in the Outer Layer of Mammalian Skin},
 journal = {Biophysical Journal},
 volume = {118},
@@ -229,10 +233,10 @@ year = {2020},
 issn = {0006-3495},
 doi = {https://doi.org/10.1016/j.bpj.2020.01.044},
 url = {https://www.sciencedirect.com/science/article/pii/S0006349520301211},
-author = {Christopher M. MacDermaid and Kyle Wm. Hall and Russell H. DeVane and Michael L. Klein and Giacomo Fiorin},
+}
 @article{Bouwstra2001,
-title={Phase behavior of lipid mixtures based on human ceramides: coexistence of crystalline and liquid phases},
 author={Bouwstra, Joke A and Gooris, Gert S and Dubbelaar, FER and Ponec, Maria},
+title={Phase behavior of lipid mixtures based on human ceramides: coexistence of crystalline and liquid phases},
 journal={Journal of lipid research},
 volume={42},
 number={11},
