### 2023-03-19 gpantel/MolPainter#18 @gpantel
We have consolidated the documentation into docs/README.md and demonstration of the usage is in docs/tutorial/README.md. Documentation has been moved out of README.md in root, which instead links to these other two markdown documents.
