+++ joss_manuscript/paper.bib
@@ -159,20 +159,6 @@ url = {http://www.ncbi.nlm.nih.gov/entrez/query.fcgi?cmd=Retrieve&db=PubMed&dopt
 volume = {82},
 year = {2002}
 }
-@article{Pantelopulos2018a,
-author = {Pantelopulos, George A and Straub, John E},
-doi = {10.1016/j.bpj.2018.10.011},
-issn = {00063495},
-journal = {Biophysical Journal},
-month = {dec},
-number = {11},
-pages = {2167--2178},
-publisher = {Biophysical Society},
-title = {{Regimes of Complex Lipid Bilayer Phases Induced by Cholesterol Concentration in MD Simulation}},
-url = {https://doi.org/10.1016/j.bpj.2018.10.011 https://linkinghub.elsevier.com/retrieve/pii/S0006349518311585},
-volume = {115},
-year = {2018}
-}
 @article{Bandara2017,
 author = {Bandara, Asanga and Panahi, Afra and Pantelopulos, George A. and Straub, John E.},
 doi = {10.1002/jcc.24516},
@@ -198,6 +184,20 @@ title = {{Critical size dependence of domain formation observed in coarse-graine
 url = {http://aip.scitation.org/doi/10.1063/1.4999709},
 volume = {147},
 year = {2017}
+}
+@article{Pantelopulos2018a,
+author = {Pantelopulos, George A and Straub, John E},
+doi = {10.1016/j.bpj.2018.10.011},
+issn = {00063495},
+journal = {Biophysical Journal},
+month = {dec},
+number = {11},
+pages = {2167--2178},
+publisher = {Biophysical Society},
+title = {{Regimes of Complex Lipid Bilayer Phases Induced by Cholesterol Concentration in MD Simulation}},
+url = {https://doi.org/10.1016/j.bpj.2018.10.011 https://linkinghub.elsevier.com/retrieve/pii/S0006349518311585},
+volume = {115},
+year = {2018}
 }
 @article{Pantelopulos2022,
 author = {Pantelopulos, George A and Matsuoka, Daisuke and Hutchison, James M and Sanders, Charles R and Sugita, Yuji and Straub, John E and Thirumalai, D},
