+++ joss_manuscript/paper.bib
@@ -172,7 +172,7 @@ url = {http://doi.wiley.com/10.1002/jcc.24516},
 volume = {38},
 year = {2017}
 }
-@article{Pantelopulos2017,
+@article{Pantel2017,
 author = {Pantelopulos, George A. and Nagai, Tetsuro and Bandara, Asanga and Panahi, Afra and Straub, John E.},
 doi = {10.1063/1.4999709},
 issn = {00219606},
@@ -185,7 +185,7 @@ url = {http://aip.scitation.org/doi/10.1063/1.4999709},
 volume = {147},
 year = {2017}
 }
-@article{Pantelopulos2018a,
+@article{Pantel2018a,
 author = {Pantelopulos, George A and Straub, John E},
 doi = {10.1016/j.bpj.2018.10.011},
 issn = {00063495},
@@ -199,7 +199,7 @@ url = {https://doi.org/10.1016/j.bpj.2018.10.011 https://linkinghub.elsevier.com
 volume = {115},
 year = {2018}
 }
-@article{Pantelopulos2022,
+@article{Pantel2022,
 author = {Pantelopulos, George A and Matsuoka, Daisuke and Hutchison, James M and Sanders, Charles R and Sugita, Yuji and Straub, John E and Thirumalai, D},
 doi = {10.1073/pnas.2212207119},
 issn = {0027-8424},
