### 2023-03-19 3842d43
Making DOCS directory, changing paths
