+++ docs/tutorial/solvate_above_membrane.toml
@@ -1,8 +1,8 @@
-solute = "tutorial/solute.pdb"
-output = "tutorial/solute+uppersolvent.pdb"
+solute = '././PDBfiles/solute.pdb'
+output = '././PDBfiles/solute+uppersolvent.pdb'
 
 [solute_lattice]
-shape = "hex" # "hex" or "square" solute lattice shape, corresponding to the MolPainter lattice
+shape = 'hex' # "hex" or "square" solute lattice shape, corresponding to the MolPainter lattice
 width = 32 # number of cells along x-axis
 height = 16 # number of cells along y-axis
 spacing = 8 # length of each cell in angstroms
@@ -14,7 +14,7 @@ lower_z_position = 16 # lower z-limit of the solvent lattice
 upper_z_position = 55 # upper z-limit of the solvent lattice
 
 [solvent_molecules]
-paths = ['tutorial/PDBfiles/NA+.pdb', 'tutorial/PDBfiles/CL-.pdb', 'tutorial/PDBfiles/WF.pdb', 'tutorial/PDBfiles/W.pdb'] # paths to pdb files of each solvent molecule
+paths = ['./PDBfiles/NA+.pdb', './PDBfiles/CL-.pdb', './PDBfiles/WF.pdb', './PDBfiles/W.pdb'] # paths to pdb files of each solvent molecule
 numbers = [54, 54, 501, 4511] # number of each molecule in paths to add
 max_iterations = 1000 # maximum number of attempts to re-insert each molecule before quitting
 rotate = false # true or false, randomly rotates solvent molecules in 3D
