### 2023-03-19 51aca20
Repaired paths for images
