+++ docs/README.md
@@ -5,7 +5,7 @@ MolPainter is a novel GUI tool for constructing layered molecular systems repres
 The purpose of MolPainter is to enable the construction of complex multilayered molecular systems through fine control over the position of individual molecules, using the allegory of painting as a means by which to select these positions in cartesian coordinates.
 
 ## MolPainter objects and functions
-![Figure 1. MolPainter components and functions. (A) *n* layers defined at assigned z-axial coordinates in angstrom units. (B) the grid used to define the xy-plane position of each lipid in terms of a cell spacing in angstrom units and number of cells in the x- and y-dimension. (C) The molecules are defined by PDB files supplied by the user. (D) Anchor points, defined in PDB files using B-factors of 1.00, define where the centroid of all anchor points z-axially align into the grid layer pained upon. (E) The canvas holds labels of each molecule painted onto xy-positions in each grid on each layer via various painting tools, such as the spray can. (G) Solutes can be inserted to the canvas, such as the C99 protein monomer shown here,[@Pantelopulos2022] obfuscating cells across multiple grid layers to prevent clashes with painted molecules.\label{fig:MolPainterFigure}](images/MolPainterObjectsAndFunctions.png)
+![Figure 1. MolPainter components and functions. (A) *n* layers defined at assigned z-axial coordinates in angstrom units. (B) the grid used to define the xy-plane position of each lipid in terms of a cell spacing in angstrom units and number of cells in the x- and y-dimension. (C) The molecules are defined by PDB files supplied by the user. (D) Anchor points, defined in PDB files using B-factors of 1.00, define where the centroid of all anchor points z-axially align into the grid layer pained upon. (E) The canvas holds labels of each molecule painted onto xy-positions in each grid on each layer via various painting tools, such as the spray can. (G) Solutes can be inserted to the canvas, such as the C99 protein monomer shown here,[@Pantelopulos2022] obfuscating cells across multiple grid layers to prevent clashes with painted molecules.\label{fig:MolPainterFigure}](tutorial/images/MolPainterObjectsAndFunctions.png)
 
 ### Layers
 Systems in MolPainter are represented by one or more layers. Each layer is represented by a grid defined at a given position on the Z-axis (Figure 1.A). This enables the creation of layered systems as well as simple three-dimensional structures.
@@ -71,7 +71,7 @@ MolSolvator solvates systems by inserting solvent molecules into a three dimensi
 
 MolSolvator inserts molecules into the solvent cells in a quantity and using topology and coordinates specified in a list of PDB file paths and a list of the number of each molecule. When MolSolvator fails to add a solvent molecule to the lattice due to overlaps with obstructed neighboring lattice sites, it will fail and attempt to add the molecule to another randomly selected lattice site until reaching a user-specified maximum number of failures. Solvent molecules can be randomly rotated in three dimensions upon each insertion attempt, both to achieve an anisotropic mixture of solvent molecules and to succeed at finding orientations of large molecules that fit into the solvent box. It is recommended that larger solvents be inserted first such that difficulties in finding grids and rotations of each inserted solvent are avoided.
 
-![First of two MolSolvator input scripts from the tutorial. Solvates a rectangular prism above the lipid bilayer constructed in the MolPainter tutorial, including ions.\label{fig:MolSolvatorFigure}](images/MolSolvatorFunctionsAndUsage.png)
+![First of two MolSolvator input scripts from the tutorial. Solvates a rectangular prism above the lipid bilayer constructed in the MolPainter tutorial, including ions.\label{fig:MolSolvatorFigure}](tutorial/images/MolSolvatorFunctionsAndUsage.png)
 
 ### MolSolvator Input Parameters
 
