+++ MolSolvator/MolSolvator.py
@@ -3,8 +3,172 @@ import numpy as np
 from scipy.spatial import distance
 import toml 
 import argparse
+import warnings
+
+def rand_rotation_matrix():
+    """
+    This is a pythonic implementation of the random 3D rotation method described
+    in the Graphics Gems book series by Jim Arvo in 1991
+    http://www.realtimerendering.com/resources/GraphicsGems/gemsiii/rand_rotation.c
+    """
+    randnums = np.random.uniform(size=(3,))
+    theta, phi, z = randnums
+    theta = theta*2.0*np.pi  # Rotation about the pole (Z).
+    phi   = phi*2.0*np.pi  # For direction of pole deflection.
+    z     = z*2.0  # For magnitude of pole deflection.
+
+    r = np.sqrt(z)
+    V = (np.sin(phi) * r,
+         np.cos(phi) * r,
+         np.sqrt(2.0 - z))
+    st = np.sin(theta)
+    ct = np.cos(theta)
+    R = np.array(((ct, st, 0), (-st, ct, 0), (0, 0, 1)))
+    M = (np.outer(V, V) - np.eye(3)).dot(R)
+    return M
+
+def create_solvent_universe(path, solvent_molecule_id, n_residues):
+    """
+    This creates and populates the universe topology for the solvent molecules
+    """
+    solvu = mda.Universe(path)
+    solvuxyz = np.copy(solvu.atoms.positions)
+    solvent_cog = np.mean(solvuxyz, axis=0)
+    solvuxyz = solvuxyz - solvent_cog
+
+    mol_n_atoms = solvu.atoms.n_atoms
+    n_atoms = mol_n_atoms*n_residues
+    resindices = np.repeat(range(n_residues), mol_n_atoms)
+    segindices = [0]*n_residues
+
+    # initialize an empty universe and populate it
+    solvent_universe = mda.Universe.empty(n_atoms, n_residues=n_residues, atom_resindex=resindices,
+                                          residue_segindex=segindices, trajectory=True)
+    solvent_universe.add_TopologyAttr('name', list(solvu.atoms.names)*n_residues)
+    solvent_universe.add_TopologyAttr('type', list(solvu.atoms.types)*n_residues)
+    solvent_universe.add_TopologyAttr('resnames', list(solvu.atoms.residues.resnames)*n_residues)
+    solvent_universe.add_TopologyAttr('resid', list(range(1, (n_residues)+1)))
+    solvent_universe.add_TopologyAttr('segid', [solvu.atoms.segids[0]])
+    solvent_coordinates = np.concatenate(np.array([solvu.atoms.positions]*n_residues))
+    return solvent_universe
+
+def lookup_lattice_xyz(lattice_xyz, lattice_sites, site_index):
+    """
+    This function returns the x, y, z coordinates of lattice_xyz from lattice_sites given the site index
+    """
+    x, y, z = lattice_xyz[lattice_sites[0][site_index]][lattice_sites[1][site_index]][lattice_sites[2][site_index]]
+    return x, y, z
+
+def assign_lattice_positions(path, solvent_lattice, solvent_lattice_xyz_coordinates, solvent_molecule_id, rotation_matrices, rotation_boolean):
+    """
+    This function sets the x, y, z position of each solvent atom in the universe
+    """
+    solvu = mda.Universe(path)
+    solvxyz = solvu.atoms.positions
+    solvcom = solvu.atoms.center_of_mass()
+
+    solvent_lattice_sites = np.array(np.where(solvent_lattice == solvent_molecule_id))
+    # assign coordinates
+    solvent_coordinates = []
+    for i in range(len(solvent_lattice_sites[0])):
+        xyz = np.copy(solvxyz - solvcom) # centered coordinates
+        if rotation_boolean == True:
[…]
