### 2023-03-19 e2a2c75
Adding even more exceptions for MolSolvator
