### 2023-03-19 ff648db
Adding version checking to MolSolvator
