### 2023-03-19 gpantel/MolPainter#20 @gpantel
Thank you for the critical comments! We have updated MolSolvator in several ways.

1) All functions are general now and have a doc string.
2) Spurious MDAnalysis warnings related to topology have been suppressed.
3) The NumPy warning was coming from a NaN that was due to using B-factors for determining the center of geometry of solvent molecules. This was not explained in the documentation or tutorial and we feel it's an unnecessary and confusing function so we removed it -- centers of geometry for placing solvents into the lattice are determined normally now.
4) MolSolvator is now a package that is part of MolPainter such that we can have it directly share the version number with MolPainter. The  molsolvator command line options -v and --version will return the MolPainter version now.
