### 2023-03-19 gpantel/MolPainter#19 @gpantel
Thank you for the helpful suggestions!

1) We have added an "About" button with important links to the software and documentation.
2) We have renamed "Open Painting" to "Load Painting".
3) Now you can click or hold to use the pencil tool.
4) The mol% is a volume-independent measure of the relative quantity of chemical species in a system. Typically in lipid bilayers, which is the main application for MolPainter, mol% is used to measure the relative amount of lipids and other molecules in a layer, so we'll keep it as-is.
