### 2023-03-19 gpantel/MolPainter#17 @gpantel
Thank you for the important suggestion.

To the README in root we have added the list of dependencies and supported platforms.

We have added the suggested About button to MolPainter. We have also added a -v and --version command line options for MolSolvator. We will include git tags.
