### 2023-03-19 gpantel/MolPainter#16 @gpantel
Thank you for pointing this out. We have now supplied contribution guidelines in CONTRIBUTIONS.md in root.
