### 2023-03-19 gpantel/MolPainter#21 @gpantel
Thank you for reporting this.

This is a known issue on Mac OS Monterey (https://docs.python.org/3.10/whatsnew/changelog.html#id71). The solution is to update to python 3.10 or higher.
