### 2023-03-19 398c4cd
Exporter now correctly reads and writes chain IDs
