### 2023-03-19 46072fe
Adding chainIDs and seed options to MolSolvator
