+++ MolPainter/MolSolvator.py
@@ -168,9 +168,9 @@ def insert_to_random_empty_point(path, solvent_lattice, solvent_molecule_id, sol
             raise Exception('Failed to insert to lattice after max_interations!')
         return solvent_lattice, rotation_matrix
 
-def main():
+def parse_and_initialize():
     """
-    MolSolvator entirely from here
+    This parses command line arguments
     """
     parser = argparse.ArgumentParser(description='Read in the TOML-format input file')
     parser.add_argument("-i", "--input", type=str, help='TOML-format input file')
@@ -203,21 +203,14 @@ def main():
         if os.path.isfile(inputs['solvent_molecules']['paths'][i]) == False:
             raise Exception('The input file %s does not exist'%inputs['solvent_molecules']['paths'][i])
     if 'seed' in inputs['solvent_molecules']:
-        np.random.seed(inputs['solvent_molecules']['seed'])
-
-    # 1) Load the solute to be solvated into a universe
-    # the warning messages about missing topology information are typically benign
-    with warnings.catch_warnings():
-        warnings.simplefilter("ignore")
-        u = mda.Universe(inputs['solute'])
-    layers_positions = u.atoms.positions
+        seed = inputs['solvent_molecules']['seed']
     
-    # 2) Define lattice_width and lattice_height coming from the canvas
+    # 1) Define lattice_width and lattice_height coming from the canvas
     lattice_width = inputs['solute_lattice']['width'] # e.g. 32 angstroms
     lattice_height = inputs['solute_lattice']['height'] # e.g. 16 angstroms
     lattice_spacing = inputs['solute_lattice']['spacing'] # e.g. 8 angstroms
     
-    # 3) Define solvent lattice spacing
+    # 2) Define solvent lattice spacing
     solvent_lattice_spacing = inputs['solvent_lattice']['spacing']# e.g. 5 angstroms
     solvent_lattice_buffer  = inputs['solvent_lattice']['buffer'] # e.g. 4 angstroms
     
@@ -228,16 +221,56 @@ def main():
     solvent_z2 = inputs['solvent_lattice']['upper_z_position'] # e.g. 72 angstroms
     solvent_lattice_depth = int(np.abs(solvent_z2 - solvent_z1)//solvent_lattice_spacing)
 
-    # 4) Set the boolean for rotation
+    # 3) Set the booleans for rotation, centerc, zeroz
+    # and set the number of attempted reinsertion interations
     if inputs['solvent_molecules']['rotate'] == True:
         rotation_boolean = True
     elif inputs['solvent_molecules']['rotate'] == False:
         rotation_boolean = False
     else:
         raise Exception("Set 'rotate' under 'solvent molecules' to true or false")
+    if args.centerc == True:
+        centerc_boolean = True
+    else:
+        centerc_boolean = False
+    if args.zeroz == True:
+        zeroz_boolean = True
+    else:
+        zeroz_boolean = False
     max_iterations = inputs['solvent_molecules']['max_iterations']
 
-    # 4) Make and array to contain solvent molecule IDs, where "0" is Empty, generate the lattice
+    # 4) Get list of paths to solute and solvent PDB files, number of solvents to insert, and output pdb path
+    solute_pdb_path = inputs['solute']
+    solvent_pdb_paths = []
+    num_solvents = []
+    for i in range(len(inputs['solvent_molecules']['paths'])):
+        solvent_pdb_paths.append(inputs['solvent_molecules']['paths'][i])
+        num_solvents.append(inputs['solvent_molecules']['numbers'][i])
+    output_pdb_path = inputs['output']
+
[…]
