### 2023-03-20 5f26b75
Test for loading system to MolPainter as JSON
