### 2023-03-20 gpantel/MolPainter#15 @gpantel
We have added manual tests for molpainter that can be run with python -m pytest. The test for MolPainter checks if it can properly load a JSON file. The test for MolSolvator checks if it can properly reproduce a solvent box given an initial seed for the RNG.
