### 2023-03-20 cd32b54
Removing JOSS manuscript stuff
