+++ joss_manuscript/paper.md
@@ -1,59 +1,0 @@
----
-title: 'MolPainter: A Tool for Painting and Solvating Layered Molecular Systems'
-tags:
-  - python
-  - molecularDynamics
-  - lipidBilayers
-authors:
-  - name: George A. Pantelopulos
-    orcid: 0000-0002-4373-1677
-    affiliation: "1,2"
-    corresponding: true
-  - name: Aaron Liberatore
-    affiliation: 3
-affiliations:
- - name: Laboratory of Chemical Physics, National Institute of Diabetes and Digestive and Kidney Diseases, National Institutes of Health, Bethesda, Maryland, USA
-   index: 1
- - name: Department of Chemistry, Boston University, Boston, Massachusetts, USA
-   index: 2
- - name: Independent Researcher, USA
-   index: 3
-date: 25 December 2022
-bibliography: paper.bib
-
----
-
-# Summary
-
-MolPainter presents the metaphor of painting as an abstraction for the construction of complex layered molecular systems. In particular, systems whose equilibrium states exhibit distinctive visual artifacts can be intentionally crafted through interactive, visually-oriented molecule placement. This is in contrast to non-interactive methods, which allow neither manual placement of molecules nor visual confirmation of desired states. The principal use-case is presented for modelling lipid bilayers with molecular dynamics simulation, though MolPainter may be used to construct any layered molecular system provided appropriate PDB-format files to define molecules (\autoref{fig:Figure1}).
-
-![MolPainter uses xy-plane grids at defined z-positions and allows the user to paint user-supplied PDB files into these grids to define a molecular system. MolSolvator solvates these systems after construction.\label{fig:Figure1}](figures/MolPainterGraphic.png)
-
-# Statement of Need
-
-Complexity in systems investigated using molecular dynamics (MD) and similar simulation methods employing MD force fields is increasing in terms of composition, initial condition, and system size. Due to limitations in time scales accessible even in highly coarse-grained force fields, careful selection of molecular composition and initial positions in complex systems is essential to designing useful simulations. With current tools for preparing molecular systems, it is straightforward to construct systems of complex composition but with random placement of molecules in the system. From such random initial distributions, systems would typically need very long, potentially intractable time scales to reach equilibrium. Construction of complex systems of realistic compositions, such as the *Mycoplasma genitalium* cell,[@Yu2016] are impressive, though it is difficult to construct any non-random initial positions for such systems.
-
-# Software with similar functionalities
-
[…]
