+++ joss_manuscript/paper.bib
@@ -1,246 +1,0 @@
-@article{Yu2016,
-author = {Yu, Isseki and Mori, Takaharu and Ando, Tadashi and Harada, Ryuhei and Jung, Jaewoon and Sugita, Yuji and Feig, Michael},
-doi = {10.7554/eLife.19274},
-issn = {2050-084X},
-journal = {eLife},
-month = {nov},
-pages = {1--22},
-pmid = {27801646},
-title = {{Biomolecular interactions modulate macromolecular structure and dynamics in atomistic model of a bacterial cytoplasm}},
-url = {http://elifesciences.org/lookup/doi/10.7554/eLife.19274},
-volume = {5},
-year = {2016}
-}
-@article{Martinez2009,
-author = {Mart{\'{i}}nez, L. and Andrade, R. and Birgin, E. G. and Mart{\'{i}}nez, J. M.},
-doi = {10.1002/jcc.21224},
-isbn = {1096-987X},
-issn = {01928651},
-journal = {Journal of Computational Chemistry},
-keywords = {animation,charge distribution,computational chemistry softwares,electrostatic potential,gui,lar density,molecu-,visualization},
-month = {oct},
-number = {13},
-pages = {2157--2164},
-pmid = {20607691},
-title = {{PACKMOL: A package for building initial configurations for molecular dynamics simulations}},
-url = {https://onlinelibrary.wiley.com/doi/10.1002/jcc.21224},
-volume = {30},
-year = {2009}
-}
-@article{VanDerSpoel2005,
-author = {{Van Der Spoel}, David and Lindahl, Erik and Hess, Berk and Groenhof, Gerrit and Mark, Alan E. and Berendsen, Herman J.C.},
-doi = {10.1002/jcc.20291},
-issn = {01928651},
-journal = {Journal of Computational Chemistry},
-keywords = {Free energy computation,GROMACS,Molecular dynamics,Molecular simulation software,Parallel computation},
-number = {16},
-pages = {1701--1718},
-pmid = {16211538},
-title = {{GROMACS: Fast, flexible, and free}},
-volume = {26},
-year = {2005}
-}
-@article{Jo2008,
-author = {Jo, Sunhwan and Kim, Taehoon and Iyer, Vidyashankara G. and Im, Wonpil},
-doi = {10.1002/jcc.20945},
-issn = {01928651},
-journal = {Journal of Computational Chemistry},
-keywords = {15,30,31,3d 4d transition,50,75,aj,density functional calculations,e,gw,mae,magnetic anisotropy energy,metal dimers,opc,orbital polarization correction,pacs numbers,xx},
-month = {aug},
-number = {11},
-pages = {1859--1865},
-pmid = {18478583},
-title = {{CHARMM-GUI: A web-based graphical user interface for CHARMM}},
-url = {http://www.ncbi.nlm.nih.gov/pubmed/18478583 http://doi.wiley.com/10.1002/jcc.20945},
-volume = {29},
-year = {2008}
-}
-@article{Schott-Verdugo2019,
-author = {Schott-Verdugo, Stephan and Gohlke, Holger},
-doi = {10.1021/acs.jcim.9b00269},
-issn = {1549-9596},
-journal = {Journal of Chemical Information and Modeling},
-month = {jun},
-number = {6},
-pages = {2522--2528},
-title = {{PACKMOL-Memgen: A Simple-To-Use, Generalized Workflow for Membrane-Protein–Lipid-Bilayer System Building}},
-url = {https://pubs.acs.org/doi/10.1021/acs.jcim.9b00269},
-volume = {59},
-year = {2019}
-}
-@article{Wassenaar2015,
[…]
