### 2023-03-20 af3456e
v1.1.3 Update (#22)
This update is in response to helpful critique of MolPainter by reviewers at JOSS as part of the peer review process.

A brief summary of changes:
Minor improvements to MolPainter and MolSolvator.
MolSolvator no longer uses B-factor fields for determining center of geometry.
Tests for MolPainter and MolSolvator.
Improved documentation.
