+++ CONTRIBUTING.md
@@ -1,0 +1,30 @@
+# Contributing
+
+We thank you for using our project and we'd love to hear from you.
+
+Contributions may include, for example,
+
+* New tutorials
+* Additional functions for modifying molecules after painting
+* Taking input files and generating output files in different formats
+
+# Contact
+
+You can contact us
+
+* Opening a discussion on GitHub
+* Contacting @gpantel via email
+
+# Reporting Issues
+
+* Open an issue
+* If MolPainter or MolSolvator experience an error please provide
+  ** The input PDB files you're using
+  ** The operating system you're working on
+  ** The parameters used for building the system
+** The version of MolPainter, MDAnalysis, and python
+
+# Let us promote your work using MolPainter!
+
+* Please reach out to us and let us know about the cool work you've done!
+* We would like to promote interesting systems constructed using MolPainter.

[…]
