### 2023-03-20 67e87b0
Changed about dialogue text box width from 60 to 62
