### 2023-03-20 gpantel/MolPainter#16 @HubLot
Great!

Just, the sublist in the "Reporting Issues" is miss formatted.
