### 2023-03-20 gpantel/MolPainter#19 @HubLot
Thanks for addressing the comments.
For the 4., thanks for the explanation, I think it would be useful to have this in the "Molecules" section of the README.md.
