### 2023-03-20 gpantel/MolPainter#17 @HubLot
Thanks for adding information in the README.

Now, it seems the 1.1.3 version can be updated to Pypi?
