### 2023-03-20 gpantel/MolPainter#18 @HubLot
Thanks for the improvements.
In the docs/README, the link to tutorial/README tutorial is wrong and point to the docs/README. There is also a bunch of latex commands still present in the file...
