### 2023-03-20 0d7be60
Addressing JOSS review response to documentation update.
