### 2023-03-20 gpantel/MolPainter#16 @gpantel
Thanks checking this and pointing that out. It's fixed now.
