### 2023-03-20 gpantel/MolPainter#17 @gpantel
Yep, we plan to update PyPI a bit later today.
