### 2023-03-20 gpantel/MolPainter#18 @gpantel
Thanks for checking these so quickly. I've replaced the remaining TeX with markdown.
