### 2023-03-20 gpantel/MolPainter#19 @gpantel
Thanks for the quick reply. I added (4) to the docs/README.md.
