### 2023-03-21 gpantel/MolPainter#17 @gpantel
Updated 1.1.3 on PyPI now.
