### 2023-03-27 gpantel/MolPainter#15 @HubLot
Thanks for the tests although it's quit minimalist.
In Molpainter, the exporter needs to be tested to make sure the PDB produced is correct compared to the GUI grid, same thing for the JSON save file and the solute feature.
