### 2023-03-27 gpantel/MolPainter#15 @gpantel
Yeah, we planned to add tests of that nature in 1.1.4. Thanks for suggesting specifics.
