### 2023-03-30 #5128 @HubLot
Hi,

My review is finished. The authors addressed the issued I opened. The paper has been lightened, community guidelines added and the documentation has been improved. Tests have been added although there are minimalist. Authors said they will add more for the next version.
Overall, I can recommend this software for publication.
