### 2023-04-09 ff49c00
1.1.4
Split exporter GUI from logic and added sanity check for MolPainter PDB generation based on tutorial
