+++ MolPainter/__init__.py
@@ -1,3 +1,3 @@
-__version__='1.1.3'
+__version__='1.1.4'
 __url__='https://github.com/gpantel/MolPainter'
 __tutorial_url__='https://github.com/gpantel/MolPainter/docs/tutorial/README.md'
