### 2023-04-10 e600333
Absolute paths in README for pypi landing page
