+++ joss_manuscript/paper.bib
@@ -173,7 +173,7 @@ volume = {38},
 year = {2017}
 }
 @article{Pantel2017,
-author = {Pantelopulos, George A. and Nagai, Tetsuro and Bandara, Asanga and Panahi, Afra and Straub, John E.},
+author = {Pantelopulos, G. A. and Nagai, Tetsuro and Bandara, Asanga and Panahi, Afra and Straub, John E.},
 doi = {10.1063/1.4999709},
 issn = {00219606},
 journal = {Journal of Chemical Physics},
@@ -185,8 +185,8 @@ url = {http://aip.scitation.org/doi/10.1063/1.4999709},
 volume = {147},
 year = {2017}
 }
-@article{Pantel2018a,
-author = {Pantelopulos, George A and Straub, John E},
+@article{Pantel2018,
+author = {Pantelopulos, G. A. and Straub, John E},
 doi = {10.1016/j.bpj.2018.10.011},
 issn = {00063495},
 journal = {Biophysical Journal},
@@ -200,7 +200,7 @@ volume = {115},
 year = {2018}
 }
 @article{Pantel2022,
-author = {Pantelopulos, George A and Matsuoka, Daisuke and Hutchison, James M and Sanders, Charles R and Sugita, Yuji and Straub, John E and Thirumalai, D},
+author = {Pantelopulos, G. A. and Matsuoka, Daisuke and Hutchison, James M and Sanders, Charles R and Sugita, Yuji and Straub, John E and Thirumalai, D},
 doi = {10.1073/pnas.2212207119},
 issn = {0027-8424},
 journal = {Proceedings of the National Academy of Sciences},
