+++ joss_manuscript/paper.md
@@ -42,7 +42,7 @@ Several useful tools have been made available for the construction of complex mo
 
 Lipid membranes are believed to exhibit specific, functional clusters of lipids, cholesterol, and proteins, often generally referred to as lipid rafts.[@Sych2021] Lipid rafts are characterized by the liquid ordered ($\mathrm{L_o}$) lipid phase, unique from the liquid disordered phase ($\mathrm{L_d}$) of the bulk membrane. The formation of the $\mathrm{L_o}$ lipid phase and phase separation from the $\mathrm{L_d}$ phase is an example of a long time scale phenomenon which is computationally intractable in all-atom MD simulation. However, as the phase diagrams of ternary lipid mixtures and cholesterol have been characterized, it is possible to estimate *a priori* whether the lipid bilayer should be phase separated, and it is also possible to infer the relative concentration of cholesterol in $\mathrm{L_o}$ and $\mathrm{L_d}$ phases. With an appropriate tool for the task, it is possible to model phase-separated states with a near-equilibrium initial spatial placement of lipids which can then be equilibrated using conventional all-atom MD simulation.
 
-The shape of such molecular phase separations and clusters can also be unique. The $\mathrm{L_o}$ phase at high cholesterol concentrations has been proposed to form "threads" of cholesterol, in which each cholesterol has no more and no less than two cholesterol in the first solvation shell, manifesting a maze-like pattern.[@Huang1999; @Miao2002; @Pantel2018a] This has been proposed to either be due to unfavorable penetration of water introduced by loss of protective lipid headgroups or specific cholesterol-cholesterol interactions involving the $\alpha$ and $\beta$ faces of cholesterol.[@Bandara2017] The shape and size of lipid phase separations can also vary significantly, manifesting as stripes or dots in vesicles[@Pantel2017] and molecular simulations with varying system dimensions and line tensions between the $\mathrm{L_o}$ and $\mathrm{L_d}$ phase.[@Kwon2022] 
+The shape of such molecular phase separations and clusters can also be unique. The $\mathrm{L_o}$ phase at high cholesterol concentrations has been proposed to form "threads" of cholesterol, in which each cholesterol has no more and no less than two cholesterol in the first solvation shell, manifesting a maze-like pattern.[@Huang1999; @Miao2002; @Pantel2018] This has been proposed to either be due to unfavorable penetration of water introduced by loss of protective lipid headgroups or specific cholesterol-cholesterol interactions involving the $\alpha$ and $\beta$ faces of cholesterol.[@Bandara2017] The shape and size of lipid phase separations can also vary significantly, manifesting as stripes or dots in vesicles[@Pantel2017] and molecular simulations with varying system dimensions and line tensions between the $\mathrm{L_o}$ and $\mathrm{L_d}$ phase.[@Kwon2022] 
 
 Along the z-axis, multi-layered lipid systems, such as the long-period-periodicity (LPP) phase of lipids,[@Bouwstra2001] have been an interesting subject of lipid biophysics for their importance in mediating permeation of small molecules through skin. Large scale molecular simulations of model multilamellar skin lipids have been performed to better understand these complex environments.[@MacDermaid2020] MolPainter is uniquely well-suited to construction of systems with complex lateral distributions as well as such multi-layered lipid systems.
 
