### 2023-04-10 #5128 @tarleb
The problem with author names is due to disambiguation: one citation uses `Georg A.` with a dot, while the others use `Georg A`. Disambiguation happens before the dot is added, so the algorithm detects them as different and adds the first name in an attempt to make the names distinct. This might be a bug in [jgm/citeproc](https://github.com/jgm/citeproc), but I'm not sure.

Either way, removing the dot after `Georg A.` in Pantel2017 fixes the issue.
