### 2023-04-17 #5128 @jgostick
Hi @gpantel, now that we have solved the bibtex issues, let's move on the process.  I have created a checkliss above with some tasks for you.  Start tackling those when you have a chance.
