### 2023-04-17 e5b7b67
Fix DOI of MacDermaid2020 bib entry
