+++ joss_manuscript/paper.bib
@@ -231,7 +231,7 @@ number = {7},
 pages = {1588-1601},
 year = {2020},
 issn = {0006-3495},
-doi = {https://doi.org/10.1016/j.bpj.2020.01.044},
+doi = {10.1016/j.bpj.2020.01.044},
 url = {https://www.sciencedirect.com/science/article/pii/S0006349520301211},
 }
 @article{Bouwstra2001,
