### 2023-04-17 #5128 @gpantel
@jgostick Ok I've corrected the DOI entry for MacDermaid2022 ( https://doi.org/10.1016/j.bpj.2020.01.044 ->  10.1016/j.bpj.2020.01.044). The entry "10.1016/s0022-2275(20)31502-9" is correct. My name is corrected in the Zenodo entry now (thanks for checking!).

Also, not important to this process, but we added the more complete test for MolPainter @HubLot requested in v1.1.4 and a bug fix and documentation improvement involving MolSolvator in v1.1.5, part of the latest tag and v1.1.5 on pypi.
