### 2023-04-19 #5128 @jgostick
Hi @gpantel, can you also add the suggested DOI for the Bouwstra paper? I checked and it seems to be correct.

Also, I need to check with the team about which version of the code should be archived on zenodo because you have 1.1.3 over there, and 1.1.5 is your latest version.  I think there is a way to have zenodo create a new doi for each release.  That is what happens for my repos (https://zenodo.org/record/7016049).
