+++ joss_manuscript/paper.bib
@@ -243,4 +243,6 @@ number={11},
 pages={1759--1770},
 year={2001},
 publisher={ASBMB}
+doi={10.1016/s0022-2275(20)31502-9},
+url={https://www.sciencedirect.com/science/article/pii/S0022227520315029},
 }
