+++ joss_manuscript/paper.bib
@@ -242,7 +242,6 @@ volume={42},
 number={11},
 pages={1759--1770},
 year={2001},
-publisher={ASBMB}
-doi={10.1016/s0022-2275(20)31502-9},
+doi={10.1016/S0022-2275(20)31502-9},
 url={https://www.sciencedirect.com/science/article/pii/S0022227520315029},
 }
