### 2023-04-19 #5128 @gpantel
Hi @jgostick,

Thanks again, the entry for Bouwstra is fixed now.

I manually updated the zenodo to the current version v1.1.5 https://doi.org/10.5281/zenodo.7846421

Next time we work on an update as a candidate for tagging we'll discuss automating zenodo and pypi updates.
