### 2023-04-19 #5128 @tarleb
To explain the error: it seems that the [`isbn` entry](https://github.com/gpantel/MolPainter/blob/paper_joss/joss_manuscript/paper.bib#L17) in Martinetz2009 contains the ISSN of the journal's web version instead of an ISBN. It's probably best to just remove the line.

Likewise, Wassenafr2015 [also](https://github.com/gpantel/MolPainter/blob/paper_joss/joss_manuscript/paper.bib#L74) has an ISSN in the ISBN field. The line should be removed.
