### 2023-04-20 def0605
Cleaning BIB entries
Removing double brackets in titles and trying sentence case.
