+++ joss_manuscript/paper.bib
@@ -6,7 +6,7 @@ journal = {eLife},
 month = {nov},
 pages = {1--22},
 pmid = {27801646},
-title = {{Biomolecular interactions modulate macromolecular structure and dynamics in atomistic model of a bacterial cytoplasm}},
+title = {Biomolecular interactions modulate macromolecular structure and dynamics in atomistic model of a bacterial cytoplasm},
 url = {http://elifesciences.org/lookup/doi/10.7554/eLife.19274},
 volume = {5},
 year = {2016}
@@ -14,7 +14,6 @@ year = {2016}
 @article{Martinez2009,
 author = {Mart{\'{i}}nez, L. and Andrade, R. and Birgin, E. G. and Mart{\'{i}}nez, J. M.},
 doi = {10.1002/jcc.21224},
-isbn = {1096-987X},
 issn = {01928651},
 journal = {Journal of Computational Chemistry},
 keywords = {animation,charge distribution,computational chemistry softwares,electrostatic potential,gui,lar density,molecu-,visualization},
@@ -22,7 +21,7 @@ month = {oct},
 number = {13},
 pages = {2157--2164},
 pmid = {20607691},
-title = {{PACKMOL: A package for building initial configurations for molecular dynamics simulations}},
+title = {PACKMOL: A package for building initial configurations for molecular dynamics simulations},
 url = {https://onlinelibrary.wiley.com/doi/10.1002/jcc.21224},
 volume = {30},
 year = {2009}
@@ -36,7 +35,7 @@ keywords = {Free energy computation,GROMACS,Molecular dynamics,Molecular simulat
 number = {16},
 pages = {1701--1718},
 pmid = {16211538},
-title = {{GROMACS: Fast, flexible, and free}},
+title = {GROMACS: Fast, flexible, and free},
 volume = {26},
 year = {2005}
 }
@@ -50,7 +49,7 @@ month = {aug},
 number = {11},
 pages = {1859--1865},
 pmid = {18478583},
-title = {{CHARMM-GUI: A web-based graphical user interface for CHARMM}},
+title = {CHARMM-GUI: A web-based graphical user interface for CHARMM},
 url = {http://www.ncbi.nlm.nih.gov/pubmed/18478583 http://doi.wiley.com/10.1002/jcc.20945},
 volume = {29},
 year = {2008}
@@ -63,7 +62,7 @@ journal = {Journal of Chemical Information and Modeling},
 month = {jun},
 number = {6},
 pages = {2522--2528},
-title = {{PACKMOL-Memgen: A Simple-To-Use, Generalized Workflow for Membrane-Protein–Lipid-Bilayer System Building}},
+title = {PACKMOL-Memgen: A simple-to-use, generalized workflow for membrane-protein–lipid-bilayer system building},
 url = {https://pubs.acs.org/doi/10.1021/acs.jcim.9b00269},
 volume = {59},
 year = {2019}
@@ -71,14 +70,13 @@ year = {2019}
 @article{Wassenaar2015,
 author = {Wassenaar, Tsjerk A. and Ing{\'{o}}lfsson, Helgi I. and B{\"{o}}ckmann, Rainer A. and Tieleman, D. Peter and Marrink, Siewert J.},
 doi = {10.1021/acs.jctc.5b00209},
-isbn = {1549-9618},
 issn = {1549-9618},
 journal = {Journal of Chemical Theory and Computation},
 month = {may},
 number = {5},
 pages = {2144--2155},
 pmid = {26574417},
-title = {{Computational Lipidomics with insane : A Versatile Tool for Generating Custom Membranes for Molecular Simulations}},
[…]
