+++ joss_manuscript/paper.bib
@@ -6,7 +6,7 @@ journal = {eLife},
 month = {nov},
 pages = {1--22},
 pmid = {27801646},
-title = {Biomolecular interactions modulate macromolecular structure and dynamics in atomistic model of a bacterial cytoplasm},
+title = {Biomolecular Interactions Modulate Macromolecular Structure and Dynamics in Atomistic Model of a Bacterial Cytoplasm},
 url = {http://elifesciences.org/lookup/doi/10.7554/eLife.19274},
 volume = {5},
 year = {2016}
@@ -21,7 +21,7 @@ month = {oct},
 number = {13},
 pages = {2157--2164},
 pmid = {20607691},
-title = {PACKMOL: A package for building initial configurations for molecular dynamics simulations},
+title = {Packmol: A Package for Building Initial Configurations for Molecular Dynamics Simulations},
 url = {https://onlinelibrary.wiley.com/doi/10.1002/jcc.21224},
 volume = {30},
 year = {2009}
@@ -35,7 +35,7 @@ keywords = {Free energy computation,GROMACS,Molecular dynamics,Molecular simulat
 number = {16},
 pages = {1701--1718},
 pmid = {16211538},
-title = {GROMACS: Fast, flexible, and free},
+title = {GROMACS: Fast, Flexible, and Free},
 volume = {26},
 year = {2005}
 }
@@ -49,7 +49,7 @@ month = {aug},
 number = {11},
 pages = {1859--1865},
 pmid = {18478583},
-title = {CHARMM-GUI: A web-based graphical user interface for CHARMM},
+title = {CHARMM-GUI: A Web-Based Graphical User Interface for CHARMM},
 url = {http://www.ncbi.nlm.nih.gov/pubmed/18478583 http://doi.wiley.com/10.1002/jcc.20945},
 volume = {29},
 year = {2008}
@@ -62,7 +62,7 @@ journal = {Journal of Chemical Information and Modeling},
 month = {jun},
 number = {6},
 pages = {2522--2528},
-title = {PACKMOL-Memgen: A simple-to-use, generalized workflow for membrane-protein–lipid-bilayer system building},
+title = {PACKMOL-Memgen: A Simple-to-Use, Generalized Workflow for Membrane-Protein–Lipid-Bilayer System Building},
 url = {https://pubs.acs.org/doi/10.1021/acs.jcim.9b00269},
 volume = {59},
 year = {2019}
@@ -76,7 +76,7 @@ month = {may},
 number = {5},
 pages = {2144--2155},
 pmid = {26574417},
-title = {Computational lipidomics with insane: A versatile tool for generating custom membranes for molecular simulations},
+title = {Computational Lipidomics With Insane: A Versatile Tool for Generating Custom Membranes for Molecular Simulations},
 url = {http://pubs.acs.org/doi/abs/10.1021/acs.jctc.5b00209},
 volume = {11},
 year = {2015}
@@ -88,7 +88,7 @@ issn = {14602059},
 journal = {Bioinformatics},
 number = {17},
 pages = {2897--2899},
-title = {MemGen: A general web server for the setup of lipid membrane simulation systems},
+title = {MemGen: A General Web Server for the Setup of Lipid Membrane Simulation Systems},
 volume = {31},
 year = {2015}
 }
@@ -100,7 +100,7 @@ journal = {Journal of Chemical Theory and Computation},
 month = {dec},
 number = {12},
 pages = {6642--6652},
-title = {BUMPy: A model-independent tool for constructing lipid bilayers of varying curvature and composition},
+title = {BUMPy: A Model-Independent Tool for Constructing Lipid Bilayers of Varying Curvature and Composition},
 url = {https://pubs.acs.org/doi/10.1021/acs.jctc.8b00765},
[…]
