### 2023-04-20 #5128 @gpantel
@tarleb Thank you very much for the detailed explanation. I changed the BIB file. I removed the isbn entries, removed the second brackets on the titles, and changed all titles to title-case. The titles were generated in sentence-case in the resulting PDF.
