### 2023-04-20 #5128 @kyleniemeyer
Hi @gpantel, I'm doing some final checks before publishing your submission.

In the paper, could you go through and make sure that citations appear before punctuation in sentences, and also have a space between them and any preceding words? It looks like they were inserted like footnotes, but these are in-text citations.
