+++ joss_manuscript/paper.md
@@ -31,29 +31,29 @@ MolPainter presents the metaphor of painting as an abstraction for the construct
 
 # Statement of Need
 
-Complexity in systems investigated using molecular dynamics (MD) and similar simulation methods employing MD force fields is increasing in terms of composition, initial condition, and system size. Due to limitations in time scales accessible even in highly coarse-grained force fields, careful selection of molecular composition and initial positions in complex systems is essential to designing useful simulations. With current tools for preparing molecular systems, it is straightforward to construct systems of complex composition but with random placement of molecules in the system. From such random initial distributions, systems would typically need very long, potentially intractable time scales to reach equilibrium. Construction of complex systems of realistic compositions, such as the *Mycoplasma genitalium* cell,[@Yu2016] are impressive, though it is difficult to construct any non-random initial positions for such systems.
+Complexity in systems investigated using molecular dynamics (MD) and similar simulation methods employing MD force fields is increasing in terms of composition, initial condition, and system size. Due to limitations in time scales accessible even in highly coarse-grained force fields, careful selection of molecular composition and initial positions in complex systems is essential to designing useful simulations. With current tools for preparing molecular systems, it is straightforward to construct systems of complex composition but with random placement of molecules in the system. From such random initial distributions, systems would typically need very long, potentially intractable time scales to reach equilibrium. Construction of complex systems of realistic compositions, such as the *Mycoplasma genitalium* cell [@Yu2016], are impressive, though it is difficult to construct any non-random initial positions for such systems.
 
 # Software with similar functionalities
 
-Several useful tools have been made available for the construction of complex molecular systems. Tools for off-lattice random insertion of molecules such as PACKMOL[@Martinez2009] and the insert-molecules program in GROMACS,[@VanDerSpoel2005] enable construction of off-lattice molecular systems by placement of randomly translated and rotated molecules into user-defined volumes with optional restraints on specific atom positions. Tools for phase-specific preparation of simulations such as CHARMM-GUI,[@Jo2008] PACKMOL-MemGen,[@Schott-Verdugo2019] and insane.py[@Wassenaar2015] for all-atom and coarse-grained lipid bilayer simulations are popular and widely used, and handle complex lipid mixtures by allowing the user to control the number and identity of lipids of lipids randomly, laterally placed within two flat leaflets. These programs also provide support for the inclusion of solutes, such as transmembrane proteins. Similar tools for constructing arbitrary, flat bilayers such as MemGen[@Knight2015], vesicles with insane.py and CHARMM-GUI, and more complex curved geometries with BUMPy[@Boyd2018] and LipidWrapper[@Durrant2014], which also insert lipids and other molecules in a random, lateral distribution, provide for more flexibility and creativity in system construction though with less complete support in preparing the system for simulation. However, there is a lack of tools to assist in the fine-tuning of the spatial placement of molecules in lipid bilayers, and more generally into any layered geometry.
[…]
