### 2023-04-21 #5128 @gpantel
@kyleniemeyer @jgostick Thank you! All citations now always appear before punctuation and always preceded by a single space.
